Update server.js
Looking at the error logs, I can see the problem:
npm error 404 @402hub/sdk@1.0.0 is not in this registry.
The issue: 402hub-api/server.js is trying to import @402hub/sdk, but that package isn't published to npm yet! It only exists in local code. We need to modify server.js to import the SDK code directly from repo instead of trying to get it from npm.